refactor: move the Phase 2 recovery primitives into node/coherence
Move AbandonChainTo and PurgeOrphanedBlockIndexEntries from main.{h,cpp}
into node/coherence.{h,cpp}, their documented home alongside the rest of
the #2865 startup coherence-recovery logic that already calls them
(issue #3125, workstream C4). Pure code move, no behavior change.

coherence.cpp gains ad-hoc externs for g_chain_trust (defined in
main.cpp, same pattern as its existing g_seen_stakes extern) and the
gridcoin/staking/chain_trust.h + node/chainman.h includes the moved code
needs (chainman.h for UpdateSyncTime). The functions sit outside
namespace GRC because the chain globals they mutate are global-namespace.
The stale "lives in main.cpp so it can use the file-local g_chain_trust"
comment in RewindToConsistentTip is updated -- since A6/A4 those
dependencies are extern-able, which is what un-strands these functions.

main.cpp -155 lines.
